fix types generation
